requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=-
#
alembic==1.12.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via passlib
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via uvicorn
cryptography==41.0.5
    # via python-jose
dnspython==2.4.2
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==2.1.0.post1
    # via fastapi
exceptiongroup==1.1.3
    # via anyio
fastapi[all]==0.104.1
    # via -r requirements.in
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.1
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.25.1
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
mako==1.3.0
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
orjson==3.9.10
    # via fastapi
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-extra-types==2.1.0
    # via fastapi
pydantic-settings==2.0.3
    # via fastapi
python-dotenv==1.0.0
    # via
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via fastapi
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sqlalchemy==1.4.0
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   uvicorn
ujson==5.8.0
    # via fastapi
uvicorn[standard]==0.24.0.post1
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
